Move most important message-service code to service
