+++ service/src/main/kotlin/fi/espoo/evaka/msg/properties/SfiSoapProperties.kt
